refactor: refactor Service Locator pattern to Dependancy injection pattern
